move PreloadedState/CombinedState into types/store.ts
